requirements.txt
google-genai
python-dotenv
scenedetect[opencv] 
opencv-python  
ffmpeg-python 
faster-whisper
